requirements.txt
flask
requests
keystoneauth1
openstacksdk
requests
flask-socketio
python-terraform